requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: linux-64
_libgcc_mutex=0.1=conda_forge
_openmp_mutex=4.5=1_llvm
alsa-lib=1.2.3=h516909a_0
aom=3.3.0=h27087fc_1
argumentparser=1.2.1=pypi_0
blas=1.0=mkl
bottleneck=1.3.4=py39hce1f21e_0
brotlipy=0.7.0=py39h27cfd23_1003
bzip2=1.0.8=h7b6447c_0
c-ares=1.18.1=h7f98852_0
ca-certificates=2022.3.29=h06a4308_0
cairo=1.16.0=ha12eb4b_1010
certifi=2021.10.8=py39h06a4308_2
cffi=1.15.0=py39hd667e15_1
charset-normalizer=2.0.4=pyhd3eb1b0_0
cryptography=36.0.0=py39h9ce1e76_0
cudatoolkit=11.3.1=h2bc3f7f_2
cycler=0.11.0=pypi_0
cython=0.29.28=py39h295c915_0
dbus=1.13.6=h5008d03_3
expat=2.4.8=h27087fc_0
ffmpeg=4.4.1=hd7ab26d_2
font-ttf-dejavu-sans-mono=2.37=hab24e00_0
font-ttf-inconsolata=3.000=h77eed37_0
font-ttf-source-code-pro=2.038=h77eed37_0
font-ttf-ubuntu=0.83=hab24e00_0
fontconfig=2.13.96=h8e229c2_2
fonts-conda-ecosystem=1=0
fonts-conda-forge=1=0
fonttools=4.31.2=pypi_0
freeglut=3.2.2=h9c3ff4c_1
freetype=2.11.0=h70c0345_0
gettext=0.19.8.1=h73d1719_1008
giflib=5.2.1=h7b6447c_0
gmp=6.2.1=h2531618_2
gnutls=3.6.15=he1e5248_0
graphite2=1.3.13=h58526e2_1001
gst-plugins-base=1.20.1=hcf0ee16_1
gstreamer=1.20.1=hd4edc92_1
harfbuzz=4.2.0=h40b6f09_0
hdf5=1.12.1=nompi_h2386368_104
icu=69.1=h9c3ff4c_0
idna=3.3=pyhd3eb1b0_0
imageio=2.16.1=pypi_0
intel-openmp=2021.4.0=h06a4308_3561
jasper=2.0.33=ha77e612_0
jbig=2.1=h7f98852_2003
jpeg=9e=h7f98852_0
keyutils=1.6.1=h166bdaf_0
kiwisolver=1.4.2=pypi_0
krb5=1.19.3=h3790be6_0
lame=3.100=h7b6447c_0
lcms2=2.12=h3be6417_0
ld_impl_linux-64=2.36.1=hea4e1c9_2
lerc=3.0=h9c3ff4c_0
libblas=3.9.0=12_linux64_mkl
libcblas=3.9.0=12_linux64_mkl
libclang=13.0.1=default_hc23dcda_0
libcurl=7.82.0=h7bff187_0
libdeflate=1.10=h7f98852_0
libdrm=2.4.109=h7f98852_0
libedit=3.1.20191231=he28a2e2_2
libev=4.33=h516909a_1
libevent=2.1.10=h9b69904_4
libffi=3.4.2=h7f98852_5
libgcc-ng=11.2.0=h1d223b6_14
libgfortran-ng=11.2.0=h69a702a_14
libgfortran5=11.2.0=h5c6108e_14
libglib=2.70.2=h174f98d_4
libglu=9.0.0=he1b5a44_1001
libiconv=1.16=h516909a_0
libidn2=2.3.2=h7f8727e_0
liblapack=3.9.0=12_linux64_mkl
liblapacke=3.9.0=12_linux64_mkl
libllvm13=13.0.1=hf817b99_2
libnghttp2=1.47.0=h727a467_0
libnsl=2.0.0=h7f98852_0
libogg=1.3.4=h7f98852_1
libopencv=4.5.5=py39hb0e02d1_7
libopus=1.3.1=h7f98852_1
libpciaccess=0.16=h516909a_0
libpng=1.6.37=hbc83047_0
libpq=14.2=hd57d9b9_0
libprotobuf=3.19.4=h780b84a_0
libssh2=1.10.0=ha56f1ee_2
libstdcxx-ng=11.2.0=he4da1e4_14
libtasn1=4.16.0=h27cfd23_0
libtiff=4.3.0=h542a066_3
libunistring=0.9.10=h27cfd23_0
libuuid=2.32.1=h7f98852_1000
libuv=1.40.0=h7b6447c_0
libva=2.14.0=h7f98852_0
libvorbis=1.3.7=h9c3ff4c_0
libvpx=1.11.0=h9c3ff4c_3
libwebp=1.2.2=h55f646e_0
libwebp-base=1.2.2=h7f8727e_0
libxcb=1.13=h7f98852_1004
libxkbcommon=1.0.3=he3ba5ed_0
libxml2=2.9.12=h885dcf4_1
libzlib=1.2.11=h166bdaf_1014
llvm-openmp=13.0.1=he0ac6c6_1
lz4-c=1.9.3=h295c915_1
matplotlib=3.5.1=pypi_0
mkl=2021.4.0=h06a4308_640
mkl-service=2.4.0=py39h7f8727e_0
mkl_fft=1.3.1=py39hd3c417c_0
mkl_random=1.2.2=py39h51133e4_0
mysql-common=8.0.28=haf5c9bc_2
mysql-libs=8.0.28=h28c427c_2
ncurses=6.3=h7f8727e_2
nettle=3.7.3=hbbd107a_1
networkx=2.7.1=pypi_0
nspr=4.32=h9c3ff4c_1
nss=3.77=h2350873_0
numexpr=2.8.1=py39h6abb31d_0
numpy=1.21.2=py39h20f2e39_0
numpy-base=1.21.2=py39h79a1101_0
opencv=4.5.5=py39hf3d152e_7
openh264=2.1.1=h4ff587b_0
openssl=1.1.1n=h7f8727e_0
packaging=21.3=pyhd3eb1b0_0
pandas=1.4.1=py39h295c915_1
pcre=8.45=h9c3ff4c_0
pillow=9.0.1=py39h22f2fdc_0
pip=21.2.4=py39h06a4308_0
pixman=0.40.0=h36c2ea0_0
progress=1.6=pypi_0
pthread-stubs=0.4=h36c2ea0_1001
py-opencv=4.5.5=py39hef51801_7
pycparser=2.21=pyhd3eb1b0_0
pyopenssl=22.0.0=pyhd3eb1b0_0
pyparsing=3.0.4=pyhd3eb1b0_0
pysocks=1.7.1=py39h06a4308_0
python=3.9.12=h9a8a25e_1_cpython
python-dateutil=2.8.2=pyhd3eb1b0_0
python_abi=3.9=2_cp39
pytorch=1.11.0=py3.9_cuda11.3_cudnn8.2.0_0
pytorch-mutex=1.0=cuda
pytz=2021.3=pyhd3eb1b0_0
pywavelets=1.3.0=pypi_0
qt=5.12.9=h1304e3e_6
readline=8.1.2=h7f8727e_1
requests=2.27.1=pyhd3eb1b0_0
scikit-image=0.19.2=pypi_0
scipy=1.8.0=pypi_0
setuptools=58.0.4=py39h06a4308_0
six=1.16.0=pyhd3eb1b0_1
sqlite=3.38.2=hc218d9a_0
svt-av1=0.9.1=h27087fc_0
tifffile=2022.3.25=pypi_0
tk=8.6.12=h27826a3_0
torchaudio=0.11.0=py39_cu113
torchvision=0.12.0=py39_cu113
typing_extensions=4.1.1=pyh06a4308_0
tzdata=2022a=hda174b7_0
urllib3=1.26.8=pyhd3eb1b0_0
wheel=0.37.1=pyhd3eb1b0_0
x264=1!161.3030=h7f98852_1
x265=3.5=h924138e_3
xorg-fixesproto=5.0=h7f98852_1002
xorg-inputproto=2.3.2=h7f98852_1002
xorg-kbproto=1.0.7=h7f98852_1002
xorg-libice=1.0.10=h7f98852_0
xorg-libsm=1.2.3=hd9c2040_1000
xorg-libx11=1.7.2=h7f98852_0
xorg-libxau=1.0.9=h7f98852_0
xorg-libxdmcp=1.1.3=h7f98852_0
xorg-libxext=1.3.4=h7f98852_1
xorg-libxfixes=5.0.3=h7f98852_1004
xorg-libxi=1.7.10=h7f98852_0
xorg-libxrender=0.9.10=h7f98852_1003
xorg-renderproto=0.11.1=h7f98852_1002
xorg-xextproto=7.3.0=h7f98852_1002
xorg-xproto=7.0.31=h7f98852_1007
xz=5.2.5=h7b6447c_0
zlib=1.2.11=h166bdaf_1014
zstd=1.5.2=ha95c52a_0
